forum

+++ GestionEvenementBack/src/main/java/tn/eesprit/gestionevenementback/Repository/LogistiqueRepository.java
@@ -7,6 +7,6 @@ import tn.eesprit.gestionevenementback.Entities.Logistique;
 import java.util.List;
 
 public interface LogistiqueRepository extends JpaRepository<Logistique, Integer> {
- Logistique findByEventId(Integer id);
+ Logistique findByEventId(Long id);
 
 }
